:recycle: refactor: FriendRelationship API 응답 생성 패턴 개선 및 서비스 로직 수정 (#8) - 생성자 직접 호출 방식을 정적 팩토리 메서드(`from`)로 변경 - 엔티티 `Builder` 패턴 적용 - 서비스 레이어 로직 개선
