+++ src/main/java/com/adit/backend/domain/user/entity/FriendRelationship.java
@@ -11,6 +11,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -21,18 +22,26 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class FriendRelationship extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "from_user_id", nullable = false)
-    private User fromUser;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "from_user_id", nullable = false)
+	private User fromUser;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "to_user_id", nullable = false)
-    private User toUser;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "to_user_id", nullable = false)
+	private User toUser;
 
-    @Column(nullable = false)
-    private String status;
+	@Column(nullable = false)
+	private String status;
+
+	@Builder
+	public FriendRelationship(Long id, User fromUser, User toUser, String status) {
+		this.id = id;
+		this.fromUser = fromUser;
+		this.toUser = toUser;
+		this.status = status;
+	}
 }
